Assignment 2 Part 1 Complete-ish
Finished part 1 of the second assignment. For some reason, the output video has very few frames that aren't overlayed correctly. The issue is either in the homography matrix calculations (probably numerical instabilities), or OpenCV not getting good matches at these frames (the problematic frames are the ones in which the book is out of the camera field of view)